+++ packages/component-library/src/components/BackToTop/BackToTop.tsx
@@ -24,7 +24,13 @@ export const BackToTop = ({ FabProps, sidekickLookup }: BackToTopProps) => {
 
   return (
     <ErrorBoundary>
-      <Root visible={trigger} {...(FabProps as any)} onClick={handleClick} data-testid="BackToTop" {...sidekick(sidekickLookup)}>
+      <Root
+        visible={trigger}
+        {...(FabProps as any)}
+        onClick={handleClick}
+        data-testid="BackToTop"
+        {...sidekick(sidekickLookup)}
+      >
         <KeyboardArrowUpIcon />
       </Root>
     </ErrorBoundary>